Move main method from test folder to src folder. Add logger. Small changes on Philosopher.java.
