+++ src/main/java/Fork.java
@@ -1,9 +1,21 @@
 public class Fork {
-    public boolean taken = false;
-    public int ID;
+    private boolean taken = false;
+    private final int ID;
 
     public Fork(int ID){
         this.ID = ID;
+    }
+
+    public boolean isTaken() {
+        return taken;
+    }
+
+    public void setTaken(boolean taken) {
+        this.taken = taken;
+    }
+
+    public int getID() {
+        return ID;
     }
 
     public synchronized void take() {
